PAYARA-1143 working on Payara Server Full
