--- appserver/extras/payara-micro/payara-micro-cdi/src/main/java/fish/payara/micro/cdi/Outbound.java
@@ -1,37 +1,0 @@
-/*
-
- DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
-
- Copyright (c) 2016 Payara Foundation. All rights reserved.
-
- The contents of this file are subject to the terms of the Common Development
- and Distribution License("CDDL") (collectively, the "License").  You
- may not use this file except in compliance with the License.  You can
- obtain a copy of the License at
- https://glassfish.dev.java.net/public/CDDL+GPL_1_1.html
- or packager/legal/LICENSE.txt.  See the License for the specific
- language governing permissions and limitations under the License.
-
- When distributing the software, include this License Header Notice in each
- file and include the License file at packager/legal/LICENSE.txt.
- */
-package fish.payara.micro.cdi;
-
-import static java.lang.annotation.ElementType.FIELD;
-import static java.lang.annotation.ElementType.METHOD;
-import static java.lang.annotation.ElementType.PARAMETER;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-import javax.inject.Qualifier;
-
-/**
- * Annotation to be applied to a Cache @Inject point to define the cache configuration
- * for the Producer to configure the cache
- * @author steve
- */
-@Retention(RetentionPolicy.RUNTIME)
-@Qualifier
-@Target({METHOD, FIELD, PARAMETER})
-public @interface Outbound {
-}
