+++ appserver/extras/payara-micro/pom.xml
@@ -11,7 +11,6 @@
     <packaging>pom</packaging>
     <name>Payara Micro Parent</name>
     <modules>
-        <module>payara-micro-cdi</module>
         <module>payara-micro-core</module>
     </modules>
