+++ appserver/payara-appserver-modules/payara-micro-cdi/pom.xml
@@ -2,10 +2,11 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <groupId>fish.payara.extras.payara-micro</groupId>
-        <artifactId>payara-micro-parent</artifactId>
+        <groupId>org.glassfish.main</groupId>
+        <artifactId>payara-appserver-modules</artifactId>
         <version>4.1.1.171-SNAPSHOT</version>
     </parent>
+    <groupId>fish.payara.appserver.micro</groupId>
     <artifactId>payara-micro-cdi</artifactId>
     <packaging>glassfish-jar</packaging>
     <name>Payara Micro CDI</name>
@@ -24,9 +25,9 @@
     </developers>
     <dependencies>
         <dependency>
-            <groupId>fish.payara.extras.payara-micro</groupId>
-            <artifactId>payara-micro-core</artifactId>
-            <version>4.1.1.171-SNAPSHOT</version>
+            <groupId>fish.payara.appserver.micro</groupId>
+            <artifactId>payara-micro-service</artifactId>
+            <version>${project.version}</version>
             <type>jar</type>
         </dependency>
         <dependency>
