+++ appserver/payara-appserver-modules/payara-micro-cdi/src/main/resources/META-INF/services/javax.enterprise.inject.spi.Extension
